Move testSnapshotWithLargeSegmentFiles to base class
